requirements.txt
scikit-learn==1.1.1
numpy==1.23.0
pytorch_lightning==1.6.4
transformers==4.20.1
torch==1.11.0
pandas==1.4.3
jupyter==1.0.0
tqdm==4.64.0
requests~=2.28.0
pyyaml~=5.4.1
setuptools~=61.2.0
fasttext==0.9.2
spacy==3.3.1
nltk==3.7
DotMap==1.3.30